Refactor : .styled.jsx에서 .style.jsx로 파일명 변경
