+++ src/components/LayOut/LayOut.style.jsx
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+export const LayOutHeader = styled.header`
+  display: flex;
+  color: blueviolet;
+  justify-content: center;
+  font-size: 30px;
+  border-bottom: 1px solid black;
+`;
