+++ src/components/Nav/Nav.style.jsx
@@ -1,0 +1,12 @@
+import styled from "styled-components";
+
+const NavWrapper = styled.div`
+  display: flex;
+  justify-content: space-around;
+`;
+
+const NavItemStyle = styled.div`
+  cursor: pointer;
+`;
+
+export { NavWrapper, NavItemStyle };
